move videos in lib/jitsi
