+++ src/lib/jitsi/flow.ts
@@ -1,7 +1,7 @@
 import { By, Key, WebElement } from 'selenium-webdriver';
 
 import { startTest, TestStep } from '../../utils/tests';
-import { verifyVideoDisplayByIndex } from '../../steps/video';
+import { verifyVideoDisplayByIndex } from './video';
 import { waitSeconds } from '../../steps/time';
 import { getCurrentUrl } from '../../steps/url';
 import {
